Refactoring UserAddress to new css handles
